+++ CHANGELOG.md
@@ -1,16 +1,15 @@
 # Changelog
 
-# Unreleased
+## Unreleased
 
 Release date: YYYY-MM-DD
 
 Code freeze date: YYYY-MM-DD
 
-### Description
-
 ### Dependency Changes
 
 Upgraded:
+
 - shapely `1.8` -> `2.0` ([#80](https://github.com/CLIMADA-project/climada_petals/pull/80))
 
 ### Added
@@ -18,6 +17,8 @@ Upgraded:
 ### Changed
 
 - refactored `climada_petals.river_flood.RiverFlood.from_nc`, removing calls to `set_raster` ([#80](https://github.com/CLIMADA-project/climada_petals/pull/80))
+- several adaptations have been necessary because of the `Tag` class being removed from the CLIMADA core package:
+  [#88](https://github.com/CLIMADA-project/climada_petals/pull/88)
 
 ### Fixed